backend/requirements.txt
Flask~=3.0.3
Flask-JWT-Extended~=4.6.0
Flask-Mailman~=1.1.1
flask-marshmallow~=1.2.1
oauthlib~=3.2.2
requests~=2.32.3
ruff==0.6.2
Werkzeug~=3.0.4
itsdangerous~=2.2.0
pymongo~=4.8.0
apispec~=6.6.1
marshmallow~=3.22.0
python-dotenv~=1.0.1
pillow~=10.3.0
Flask-Cors~=5.0.0
flask-apispec~=0.11.4